Memory leak fixes, clean up of periodic matching routine, fixed build error for MPI_STATUS_IGNORE in serial, Tecplot binary serial fix for AD mode, updated some regressions.

+++ Common/src/linear_solvers_structure.cpp
@@ -783,7 +783,7 @@ unsigned long CSysSolve::Solve(CSysMatrix & Jacobian, CSysVector & LinSysRes, CS
 
     AD::StartRecording();
     
-    AD::SetExtFuncOut(&LinSysSol[0], LinSysSol.GetLocSize());
+    AD::SetExtFuncOut(&LinSysSol[0], (int)LinSysSol.GetLocSize());
    
 #ifdef CODI_REVERSE_TYPE
     AD::FuncHelper->addUserData(&LinSysRes);
